python_tests/requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
idna==2.10
iniconfig==1.1.1
packaging==20.7
pluggy==0.13.1
py==1.9.0
PyHamcrest==2.0.2
pyparsing==2.4.7
pytest==6.1.2
requests==2.25.0
toml==0.10.2
urllib3==1.26.2
